Implement a ouath client for httpx
This let use parallel requests, handy.

See https://docs.authlib.org/en/latest/client/oauth2.html
And: https://docs.authlib.org/en/latest/client/httpx.html
